Correct the name of Melbourne City Council
The actual name of “Melbourne City” is Melbourne City Council.

This updates the name of the authority and the popolo id's of the
attached councillors, which are derived from this name.

UUIDs are starting to look attractive as it might become a pain to
change the popolo id of councillors down the line.