Preserve a snapshot of the current CSV files
This is to save the current versions of all the files ahead of
normalising the headers of all the CSV files.

There's some useful data in these files like phone numbers, and the
gender of WA councillors. We want to preserve access to that, whilst
making it easier for people to work with the CSVs in use.

Part of https://github.com/openaustralia/australian_local_councillors_popolo/issues/131