Add method to request csv from remote URL
This adds some libraries to help test the remote API.

One decision here is to gitignore and expire the VCR cassettes so that
the actual API is tested pretty regularly when someone is working on
this.

Maybe with it being so specific to the PlanningAlerts API this method
should be more explicitly about PlanningAlerts? I'm just trying to do
the dumb thing here.